# openapi.yaml
openapi: 3.0.3
info:
  title: Doctor.mx API
  description: |
    Doctor.mx - Plataforma integral de telemedicina para México.
    
    Esta API proporciona acceso a todas las funcionalidades de la plataforma Doctor.mx,
    incluyendo consultas médicas virtuales, gestión de citas, recetas digitales,
    análisis de IA, consentimientos digitales, y más.
    
    ## Autenticación
    La API utiliza Supabase Auth para la autenticación. Incluye el token de sesión
    en el header `Authorization: Bearer <token>`.
    
    ## Roles
    - **patient**: Pacientes que buscan atención médica
    - **doctor**: Médicos que brindan consultas
    - **admin**: Administradores del sistema
    
    ## Rate Limiting
    Las solicitudes están limitadas por IP y usuario autenticado.
    Headers de respuesta:
    - `X-RateLimit-Limit`: Límite de solicitudes
    - `X-RateLimit-Remaining`: Solicitudes restantes
    - `X-RateLimit-Reset`: Tiempo de reinicio
  version: 1.0.0
  contact:
    name: Doctor.mx Support
    email: soporte@doctor.mx
  license:
    name: Propietario
    url: https://doctor.mx/terms
servers:
  - url: https://doctor.mx
    description: Producción
  - url: https://staging.doctor.mx
    description: Staging
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/appointments/{id}/video:
    get:
      summary: Token de video
      description: Obtiene token para sala de video de la cita
      tags:
        - Appointments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Token de video
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  roomUrl:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== PATIENTS ====================
...
  /api/ai/followup:
    post:
      summary: Generar seguimiento
      description: Genera mensajes de seguimiento post-consulta
      tags:
        - AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                consultationId:
                  type: string
                  format: uuid
                followUpType:
                  type: string
                  enum: [24h, 7d, 30d]
      responses:
        '200':
          description: Seguimiento generado
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/consent/audit:
    get:
      summary: Auditoría de consentimientos
      description: Obtiene registros de auditoría (admin)
      tags:
        - Consent
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Registros de auditoría
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de sesión de Supabase
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Cookie de sesión de Supabase
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
...
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Se requiere autenticación"
    Forbidden:
      description: Acceso denegado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Forbidden"
            message: "No tiene permisos para esta acción"
...